Roll bulk index operation back into adapters
The adapters handle bulk index operations slightly differently to
determine which index to use. Roll this back into the adapters until a
helpful abstraction becomes more clear.
