+++ lib/elasticsearch/extensions/documents.rb
@@ -6,7 +6,6 @@ require "elasticsearch/extensions/documents/document"
 require "elasticsearch/extensions/documents/index"
 require "elasticsearch/extensions/documents/aliased_index_store"
 require "elasticsearch/extensions/documents/direct_index_store"
-require "elasticsearch/extensions/documents/bulk_indexer"
 require "elasticsearch/extensions/documents/queryable"
 require "elasticsearch/extensions/documents/storage"
 require "elasticsearch/extensions/documents/utils"
