+++ lib/elasticsearch/extensions/documents/direct_index_store.rb
@@ -38,6 +38,25 @@ module Elasticsearch
           storage.create_index(Documents.index_name)
         end
 
+        def bulk_index(documents)
+          client.bulk body: bulk_index_operations(documents)
+        end
+
+        def bulk_index_operations(documents)
+          documents.collect { |document| bulk_index_operation_hash(document) }
+        end
+
+        def bulk_index_operation_hash(document)
+          {
+            index: {
+              _index: Documents.index_name,
+              _type:  document.class.type,
+              _id:    document.id,
+              data:   document.as_hash,
+            }
+          }
+        end
+
       end
     end
   end
